@@ -2,10 +2,14 @@
   "name": "visualizer-on-tabs",
   "version": "0.0.0",
   "description": "",
+  "type": "module",
+  "module": "main/index.js",
   "scripts": {
     "clean": "shx rm -rf out",
     "copy-static": "shx mkdir -p out && shx cp -r src/content/* out",
     "dev": "node bin/build.js --debug=1 --config=./dev.json && npx serve out/",
+    "prettier": "prettier --check ./",
+    "prettier-write": "prettier --write ./",
     "eslint": "eslint bin main src",
     "webpack": "node bin/build.js --debug=1",
     "watch": "npm run copy-static && npm run webpack -- --watch",
@@ -18,6 +22,7 @@
     "@babel/preset-react": "^7.24.1",
     "babel-loader": "^9.1.3",
     "bootstrap": "^5.3.3",
+    "eslint-config-zakodium": "^16.0.0",
     "fs-extra": "^7.0.1",
     "iframe-bridge": "^1.1.0",
     "js-yaml": "^3.12.1",
@@ -31,16 +36,11 @@
     "superagent": "^4.1.0",
     "webpack": "^5.91.0"
   },
-  "main": "main/index.js",
   "devDependencies": {
     "babel-eslint": "^10.0.1",
-    "eslint": "^5.12.0",
-    "eslint-config-cheminfo": "^1.20.0",
-    "eslint-config-neptune-react": "^2.6.0",
-    "eslint-plugin-import": "^2.14.0",
-    "eslint-plugin-jest": "^22.1.3",
-    "eslint-plugin-react": "^7.12.4",
+    "eslint": "^9.34.0",
     "jest": "^29.7.0",
+    "prettier": "^3.6.2",
     "shx": "^0.3.2"
   },
   "volta": {
