@@ -1,32 +1,25 @@
+/* eslint-disable no-await-in-loop */
+
 import IframeBridge from 'iframe-bridge';
 import React from 'react';
-import { Visualizer } from 'react-visualizer';
 import Tab from 'react-bootstrap/Tab';
 import BTabs from 'react-bootstrap/Tabs';
+import { Visualizer } from 'react-visualizer';
 
+import { getConfig } from '../config/config.js';
 import Tabs from '../main/Tabs';
 import iframeMessageHandler from '../main/iframeMessageHandler';
-import tabStorage from '../main/tabStorage';
+import * as tabStorage from '../main/tabStorage';
 import { rewriteURL } from '../util';
 
 import Login from './Login';
 import TabTitle from './TabTitle';
 
-const conf = require('../config/config.js');
-
-// Setting this to true should load all the tabs on page load
-// It is discouraged to do this because loading hidden iframes
-// lead to layout issues. Especially in Firefox.
-const loadHidden = conf.loadHidden || false;
-
-const possibleViews = conf.possibleViews;
-const forbiddenPossibleViews = Object.keys(possibleViews);
-
 let tabInit = Promise.resolve();
 let currentIframe;
 
 const pageURL = new URL(window.location);
-const pageQueryParameters = (function () {
+const pageQueryParameters = (() => {
   let params = {};
   for (let key of pageURL.searchParams.keys()) {
     params[key] = pageURL.searchParams.get(key);
@@ -46,10 +39,11 @@ class App extends React.Component {
       if (level2 === 'connect' && data.windowID !== undefined) {
         if (!currentIframe) {
           // The iframe was refreshed
-          possibleViews[this.state.activeTabKey].windowID = data.windowID;
+          this.config.possibleViews[this.state.activeTabKey].windowID =
+            data.windowID;
           this.sendData(this.state.activeTabKey);
         } else {
-          possibleViews[currentIframe.id].windowID = data.windowID;
+          this.config.possibleViews[currentIframe.id].windowID = data.windowID;
           currentIframe.resolve();
           currentIframe = null;
         }
@@ -72,36 +66,49 @@ class App extends React.Component {
 
     this.state = {
       viewsList: [],
-      activeTabKey: 0
+      activeTabKey: 0,
+      isConfigLoaded: false,
+      config: {},
     };
 
+    this.loadConfig().then(() => this.loadTabs());
     this.loadTabs();
   }
 
+  async loadConfig() {
+    const config = await getConfig();
+    this.setState((state) => {
+      state.config = config;
+      state.isConfigLoaded = false;
+    });
+  }
   async loadTabs() {
     let firstTab;
     const loadTab = async (view) => {
       if (!firstTab) firstTab = view.id;
       await this.doTab(view, {
         noFocus: true,
-        load: loadHidden,
-        noFocusEvent: true
+        // Setting this to true should load all the tabs on page load
+        // It is discouraged to do this because loading hidden iframes
+        // lead to layout issues. Especially in Firefox.
+        load: this.config.loadHidden,
+        noFocusEvent: true,
       });
     };
     const data = tabStorage.load();
     // Load possible views first
-    for (let key in possibleViews) {
-      possibleViews[key].id = key;
+    for (let key in this.config.possibleViews) {
+      this.config.possibleViews[key].id = key;
       let saved;
       if ((saved = data.find((el) => el.id === key))) {
         await loadTab(saved);
       } else {
-        await loadTab(possibleViews[key]);
+        await loadTab(this.config.possibleViews[key]);
       }
     }
 
     for (let i = 0; i < data.length; i++) {
-      if (!possibleViews[data[i].id]) {
+      if (!this.config.possibleViews[data[i].id]) {
         await loadTab(data[i]);
       }
     }
@@ -117,22 +124,24 @@ class App extends React.Component {
 
   setTabStatus(data) {
     // Find view with given window ID
-    const ids = Object.keys(possibleViews);
-    let id = ids.find((id) => possibleViews[id].windowID === data.windowID);
+    const ids = Object.keys(this.config.possibleViews);
+    let id = ids.find(
+      (id) => this.config.possibleViews[id].windowID === data.windowID,
+    );
     if (!id) return;
-    let view = possibleViews[id];
+    let view = this.config.possibleViews[id];
 
     view = this.state.viewsList.find((el) => el.id === view.id);
     if (!view) return;
 
-    view.status = Object.assign({}, view.status, data.message);
+    view.status = { ...view.status, ...data.message };
     this.setState((state) => ({
-      viewsList: state.viewsList
+      viewsList: state.viewsList,
     }));
   }
 
   sendTabMessage(data) {
-    const viewInfo = possibleViews[data.id];
+    const viewInfo = this.config.possibleViews[data.id];
     if (viewInfo) {
       IframeBridge.postMessage('tab.message', data.message, viewInfo.windowID);
     }
@@ -141,28 +150,31 @@ class App extends React.Component {
   async focusTab(tabId) {
     if (this.state.viewsList.find((el) => el.id === tabId)) {
       await this.showTab(tabId, {
-        noData: true
+        noData: true,
       });
     }
   }
 
   async doTab(obj, options) {
-    if (!possibleViews[obj.id]) {
-      possibleViews[obj.id] = {
+    if (!this.config.possibleViews[obj.id]) {
+      this.config.possibleViews[obj.id] = {
         id: obj.id,
         url: obj.url,
         data: obj.data,
         closable: obj.closable,
-        rawIframe: obj.rawIframe
+        rawIframe: obj.rawIframe,
       };
     } else {
-      possibleViews[obj.id].data = obj.data;
+      this.config.possibleViews[obj.id].data = obj.data;
     }
 
-    if (conf.rewriteRules) {
-      let newURL = rewriteURL(conf.rewriteRules, possibleViews[obj.id].url);
+    if (this.state.config.rewriteRules) {
+      let newURL = rewriteURL(
+        this.config.rewriteRules,
+        this.config.possibleViews[obj.id].url,
+      );
       if (newURL) {
-        possibleViews[obj.id].rewrittenUrl = newURL;
+        this.config.possibleViews[obj.id].rewrittenUrl = newURL;
       }
     }
 
@@ -177,16 +189,16 @@ class App extends React.Component {
     const focusedTabId = options.noFocus ? undefined : id;
     let viewFromList = this.state.viewsList.find((el) => el.id === id);
     const newTab = !viewFromList;
-    const viewInfo = possibleViews[id];
+    const viewInfo = this.config.possibleViews[id];
 
     if (!viewInfo) throw new Error('unreachable');
     if (!viewFromList) {
       viewFromList = {
-        id: id,
+        id,
         url: viewInfo.url,
         rewrittenUrl: viewInfo.rewrittenUrl,
         closable: viewInfo.closable,
-        rawIframe: viewInfo.rawIframe
+        rawIframe: viewInfo.rawIframe,
       };
       this.state.viewsList.push(viewFromList);
     }
@@ -197,11 +209,13 @@ class App extends React.Component {
     // We need to get the IframeBridge ID of that frame and prevent any other iframes
     // to load during that time
     if (firstRender) {
+      // TODO: there is probably a cleaner way than a global promise
+      // eslint-disable-next-line require-atomic-updates
       tabInit = new Promise((resolve) => {
         viewFromList.rendered = true;
         this.setState((state) => ({
           activeTabKey: focusedTabId,
-          viewsList: state.viewsList
+          viewsList: state.viewsList,
         }));
 
         setTimeout(() => {
@@ -212,14 +226,14 @@ class App extends React.Component {
         }, 3000);
         currentIframe = {
           resolve,
-          id
+          id,
         };
       });
       await tabInit;
     } else {
       this.setState((state) => ({
         activeTabKey: focusedTabId,
-        viewsList: state.viewsList
+        viewsList: state.viewsList,
       }));
     }
 
@@ -237,20 +251,19 @@ class App extends React.Component {
   }
 
   sendData(id) {
-    const viewInfo = possibleViews[id];
+    const viewInfo = this.config.possibleViews[id];
     IframeBridge.postMessage(
       'tab.data',
-      Object.assign({}, viewInfo.data, {
-        queryParameters: pageQueryParameters
-      }),
-      viewInfo.windowID
+      { ...viewInfo.data, queryParameters: pageQueryParameters },
+      viewInfo.windowID,
     );
   }
 
   async removeTab(id) {
+    const forbiddenPossibleViews = Object.keys(this.config.possibleViews);
     tabStorage.remove(id);
     if (forbiddenPossibleViews.indexOf(id) === -1) {
-      delete possibleViews[id];
+      delete this.config.possibleViews[id];
     }
     let idx = this.state.viewsList.findIndex((el) => el.id === id);
     if (idx === -1) return;
@@ -273,25 +286,31 @@ class App extends React.Component {
 
     await this.showTab(newActiveTab, {
       noData: true,
-      force: true
+      force: true,
     });
   }
 
   sendTabFocusEvent(key) {
-    if (possibleViews[key]) {
-      IframeBridge.postMessage('tab.focus', {}, possibleViews[key].windowID);
+    if (this.config.possibleViews[key]) {
+      IframeBridge.postMessage(
+        'tab.focus',
+        {},
+        this.config.possibleViews[key].windowID,
+      );
     }
   }
 
   async onActiveTab(key) {
     await this.showTab(key, {
-      noData: true
+      noData: true,
     });
   }
 
   render() {
     const arr = [];
-
+    if (!this.state.isConfigLoaded) {
+      return null;
+    }
     for (let view of this.state.viewsList) {
       const closable = view.closable === undefined ? true : view.closable;
       const saved =
@@ -320,9 +339,11 @@ class App extends React.Component {
               url="visualizer.html"
               viewURL={view.rewrittenUrl || view.url}
               version={
-                this.visualizerVersion || conf.visualizerVersion || 'auto'
+                this.visualizerVersion ||
+                this.config.visualizerVersion ||
+                'auto'
               }
-              config={conf.visualizerConfig}
+              config={this.config.visualizerConfig}
               style={iframeStyle}
             />
           );
@@ -344,13 +365,13 @@ class App extends React.Component {
           eventKey={view.id}
         >
           {viewPage}
-        </Tab>
+        </Tab>,
       );
     }
 
     return (
       <div className="visualizer-on-tabs-app">
-        <Login />
+        <Login config={this.state.config} />
         <div className="visualizer-on-tabs-content d-flex flex-column">
           <BTabs
             id="visualizer-on-tabs-tab"
