@@ -2,19 +2,19 @@ import { EventEmitter } from 'events';
 
 const Tabs = new EventEmitter();
 
-Tabs.openTab = function (data) {
+Tabs.openTab = function openTab(data) {
   Tabs.emit('openTab', data);
 };
 
-Tabs.status = function (data) {
+Tabs.status = function status(data) {
   Tabs.emit('status', data);
 };
 
-Tabs.sendMessage = function (message) {
+Tabs.sendMessage = function sendMessage(message) {
   Tabs.emit('message', message);
 };
 
-Tabs.focus = function (message) {
+Tabs.focus = function focus(message) {
   Tabs.emit('focus', message);
 };
 
