@@ -6,8 +6,8 @@ module.exports = {
       {
         test: /\.jsx?$/,
         exclude: ['node_modules'],
-        use: ['babel-loader']
-      }
-    ]
-  }
+        use: ['babel-loader'],
+      },
+    ],
+  },
 };
